+++ src/Farkle/Grammar/GrammarTypes.fs
@@ -183,7 +183,18 @@ type Symbols = {
     NoiseSymbols: Noise ImmutableArray
 }
 
+/// An object that provides optimized functions for some common operations on Grammars.
+/// These functions require some computationally expensive pre-processing, which is
+/// performed only once, at the creation of this object.
+type internal OptimizedOperations = {
+    _GetNextDFAState: char -> DFAState -> Indexed<DFAState> voption
+}
+with
+    /// Gets the next DFA state from the given current one, when the given character is encountered.
+    member x.GetNextDFAState c state = x._GetNextDFAState c state
+
 /// A context-free grammar according to which, Farkle can parse text.
+[<NoComparison; ReferenceEquality>]
 type Grammar = internal {
     // This field is totally informative; it serves only the template maker.
     _Properties: ImmutableDictionary<string,string>
@@ -197,6 +208,9 @@ type Grammar = internal {
     _Groups: Group SafeArray
     _LALRStates: LALRState StateTable
     _DFAStates: DFAState StateTable
+
+    // This field makes Farkle fast(er).
+    OptimizedOperations: OptimizedOperations
 }
 with
     /// Metadata about the grammar. See the [GOLD Parser's documentation for more](http://www.goldparser.org/doc/egt/index.htm).